requirements.txt
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
enum34==1.1.10
filelock==3.6.0
fonttools==4.33.3
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
miditoolkit==0.1.16
mido==1.2.10
numpy==1.21.6
packaging==21.3
pandas==1.1.5
Pillow==9.1.0
PyGuitarPro==0.6
pylab-sdk==1.3.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0
transformers==4.18.0
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
